fix: Remove "seguimiento" steps, add specificity
This change allows Nova to manage risk that haven't happened yet because
you only have to identify it while you are planning, as it is expected
to happen during the iteration planning. Also there were steps for the
monitoring of risks (now hidden as comments) that should belong to a
daily process directly tied with the plan monitoring (PMC) since it is
only then that it can be justifiable to activate either a contingency
plan or a mitigation plan.